requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
ultralytics~=8.1.34
matplotlib~=3.8.3
torchvision~=0.17.1
pillow~=10.2.0